Use size_t arg in CPDF_InteractiveForm::GetField()
Its only usage is to pass the arg into GetFieldAtIndex(), which already
takes a size_t.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/87050
Auto-Submit: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfdoc/cpdf_interactiveform.cpp
@@ -619,7 +619,7 @@ size_t CPDF_InteractiveForm::CountFields(const WideString& csFieldName) const {
 }
 
 CPDF_FormField* CPDF_InteractiveForm::GetField(
-    uint32_t index,
+    size_t index,
     const WideString& csFieldName) const {
   if (csFieldName.IsEmpty())
     return m_pFieldTree->GetRoot()->GetFieldAtIndex(index);
